pyproject.toml
[tool.poetry]
name = "discord-code-bot"
version = "0.1.0"
description = ""
authors = ["Владимир Верхов <vladimirverhov@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "^2.4.0"
python-dotenv = "^1.0.1"
psutil = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
